telnetfilter: Don't split lines on '>' character
This removes old prompt detection code that relied on the '>' character.
We have IAC GA for prompt detection and the old code was only causing
trouble.

With the XML mode, the '>' character is common in the middle of lines
and the old prompt detection caused partial lines to be sent to client
at each xml tag boundary, confusing clients.

git-svn-id: svn://svn.code.sf.net/p/mmapper/code/trunk/mmapper@224 14c008a5-6e10-0410-bf79-c100e0f1932b