Fix exits/emulated behavior
Currently, the Exits/emulated: lines are pretty buggy, and often
display information for the wrong room.  The output is printed before
the character's location is synced, so mmapper tries to guess where
the room will be based on the queued movement commands.
Unfortunately, this is pretty problematic:

 * following doesn't work, because the order MUME sends information is
    like:

Arwen Undómiel leaves east.
You follow Arwen Undómiel.
<movement dir=east/><room><name>The Courtyard</name>

m_move is set to CID_EAST on the <movement> tag, but it's enqueued the
line before, on the "You follow" line, at which point it's still
CID_LOOK.

 * Scouting doesn't work, because the queue is prepended with
   CID_SCOUT rather than CID_EAST.

 * Regular movement in the Old Forest doesn't work, because the
   prediction isn't reliable.

 * Teleporting/scrying don't work, because there's no way to guess the
   location from just the commands.

Instead, the emulated exits should be printed _after_ the pathmachine
syncs up.

This patch makes that happen in XML mode.  It doesn't fix it for
regular mode, but it shouldn't make things worse there either.