Fixed ancient bug where mudlle'd movement would break map syncing
The PathMachine will try all exits and neighboring coordinates if it gets a
<movement/> XML tag that corresponds to a CommandEnum::NONE within the
ParseEvent. However, an oversight in the Approved class meant that it would
never find a matching room because the logic would invalidate all options as
soon as more than one comparison was made.

The fix is to now only prevent syncing if more than one valid candidate room
is found. Invalid comparisons will now be ignored.