Bugfix invalid uses of std::initializer_list (UB)
A std::initializer_list<T> is actually just a view of a compiler
generated temporary array. The temporary array's lifetime is only
valid until the end of the expression that creates the list.

Therefore, it's safe to call a function with a freshly-constructed
list in a single expression, but it's unsafe to construct a list
in one expression, let the temporary array's lifetime end, and then
attempt to use the list in a second expression.