bugfix: crash on use after free
The root cause is that the QStringRef was getting re-assigned to
point to a string owned by the QRegularExpressionMatch, so the
reference became invalid when the match went out of scope and
deleted the string.

Switching from QStringRef to QString fixes the problem by making
a copy of the string owned by QRegularExpressionMatch to avoid
referring to a deleted string.

(This could also be fixed by just stepping forward in the ref where
the code was reassigning it to point to a different string.)